requirements.txt
# Core LangChain packages
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1

# HuggingFace and Transformers
transformers>=4.30.0
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector store and embeddings
docarray>=0.40.0
numpy>=1.24.0

# Environment and configuration
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0

# Document loaders - PDF support
pypdf>=3.17.0
pdfplumber>=0.10.0

# Document loaders - Word support
python-docx>=1.0.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional utilities
pydantic>=2.0.0
requests>=2.31.0
